Move storer batch handling into separate file
