+++ internal/storer/storer.go
@@ -1,8 +1,6 @@
 package storer
 
 import (
-	"fmt"
-
 	"github.com/johnstarich/go-wasm/internal/fsutil"
 	"github.com/spf13/afero"
 )
@@ -15,10 +13,6 @@ type Storer interface {
 	GetFileRecord(path string, dest *FileRecord) error
 	// SetFileRecord assigns 'data' to the given 'path'. Returns an error if the data could not be set.
 	SetFileRecord(path string, src *FileRecord) error
-}
-
-type BatchGetter interface {
-	GetFileRecords(paths []string, dest []*FileRecord) []error
 }
 
 type fileStorer struct {
@@ -47,37 +41,4 @@ func (f *fileStorer) SetFile(path string, file *fileData) error {
 		return <-QueueSetFileRecord(f.Storer, path, nil)
 	}
 	return <-QueueSetFileRecord(f.Storer, path, &file.FileRecord)
-}
-
-func (f *fileStorer) GetFiles(paths ...string) ([]*File, []error) {
-	fileRecords := make([]*FileRecord, len(paths))
-	files := make([]*File, len(paths))
-	for i := range files {
-		path := fsutil.NormalizePath(paths[i])
-		files[i] = &File{
-			fileData: &fileData{
-				path:   path,
-				storer: f,
-			},
-		}
-		fileRecords[i] = &files[i].FileRecord
-	}
-	errs := GetFileRecords(f.Storer, paths, fileRecords)
-	return files, errs
-}
-
-func GetFileRecords(s Storer, paths []string, dest []*FileRecord) []error {
-	if len(paths) != len(dest) {
-		panic(fmt.Sprintf("GetFileRecords: Paths and dest lengths must be equal: %d != %d", len(paths), len(dest)))
-	}
-	if batcher, ok := s.(BatchGetter); ok {
-		return batcher.GetFileRecords(paths, dest)
-	}
-
-	errs := make([]error, len(paths))
-	for i := range paths {
-		path, destI := paths[i], dest[i]
-		errs[i] = s.GetFileRecord(path, destI)
-	}
-	return errs
 }
